move tests/Unit/BaseTestCase to tests/Common
