+++ tests/Unit/GlobalTracerTest.php
@@ -5,6 +5,7 @@ namespace DDTrace\Tests\Unit;
 use DDTrace\GlobalTracer;
 use DDTrace\Tracer;
 use DDTrace\Transport\Noop;
+use DDTrace\Tests\Common\BaseTestCase;
 
 final class GlobalTracerTest extends BaseTestCase
 {
