+++ tests/Unit/Http/RequestTest.php
@@ -3,7 +3,7 @@
 namespace DDTrace\Tests\Unit\Http;
 
 use DDTrace\Http\Request;
-use DDTrace\Tests\Unit\BaseTestCase;
+use DDTrace\Tests\Common\BaseTestCase;
 
 final class RequestTest extends BaseTestCase
 {
