+++ tests/Unit/Integrations/DefaultIntegrationConfigurationTest.php
@@ -3,7 +3,7 @@
 namespace DDTrace\Tests\Unit\Integrations;
 
 use DDTrace\Integrations\DefaultIntegrationConfiguration;
-use DDTrace\Tests\Unit\BaseTestCase;
+use DDTrace\Tests\Common\BaseTestCase;
 
 final class DefaultIntegrationConfigurationTest extends BaseTestCase
 {
