+++ tests/Unit/Integrations/IntegrationsLoaderTest.php
@@ -5,7 +5,7 @@ namespace DDTrace\Tests\Unit\Integrations;
 use DDTrace\Configuration;
 use DDTrace\Integrations\Integration;
 use DDTrace\Integrations\IntegrationsLoader;
-use DDTrace\Tests\Unit\BaseTestCase;
+use DDTrace\Tests\Common\BaseTestCase;
 use DDTrace\Util\Versions;
 
 final class IntegrationsLoaderTest extends BaseTestCase
