+++ tests/Unit/Log/PsrLoggerTest.php
@@ -3,7 +3,7 @@
 namespace DDTrace\Tests\Unit\Log;
 
 use DDTrace\Log\PsrLogger;
-use DDTrace\Tests\Unit\BaseTestCase;
+use DDTrace\Tests\Common\BaseTestCase;
 
 class PsrLoggerTest extends BaseTestCase
 {
