+++ tests/Unit/Obfuscation/WildcardToRegexTest.php
@@ -3,7 +3,7 @@
 namespace DDTrace\Tests\Unit\Obfuscation;
 
 use DDTrace\Obfuscation\WildcardToRegex;
-use DDTrace\Tests\Unit\BaseTestCase;
+use DDTrace\Tests\Common\BaseTestCase;
 
 final class WildcardToRegexTest extends BaseTestCase
 {
