+++ tests/Unit/Util/ContainerInfoTest.php
@@ -2,7 +2,7 @@
 
 namespace DDTrace\Tests\Unit\Util;
 
-use DDTrace\Tests\Unit\BaseTestCase;
+use DDTrace\Tests\Common\BaseTestCase;
 use DDTrace\Util\ContainerInfo;
 
 final class ContainerInfoTest extends BaseTestCase
